Facial Recognition
Modified a facial recognition algorithm to fit our use case based on CNN and alexnet. Requires future work to enable easily adding just a single face to the network, and then needs to be placed instead of requesting a username be asked during login.

Modified Arduino to an untested method to prevent deletion of all admins